Initial version of dtbook-to-odt documentation
The documentation is written in org-mode and can be exported to HTML. It
has examples, each consisting of a DTBook XML snippet and the
corresponding OpenDocument Text XML. In HTML the former is shown as
code, the latter is rendered as it would look like in OpenOffice or MS
Word.

The documentation is currently in English, except for the examples which
are in German. The examples were taken from
src/test/xprocspec/test_content.xprocspec.

As a possible extension, the examples could be automatically
extracted (tangled), converted to tests and then run, in order to verify
the correctness of the documentation.